refactor: lobbyをentities layerに移動
