+++ src/_pages/lobby-creation/ui/lobby-form.tsx
@@ -4,8 +4,8 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useRouter } from "next/navigation";
 import { Controller, useForm } from "react-hook-form";
 import { createLobby } from "@/_pages/lobby-creation/api/create-lobby";
-import { creationLobySchema } from "@/features/lobby/schemas";
-import type { LobbyCreationDate } from "@/features/lobby/types";
+import type { LobbyCreationDate } from "@/entities/lobby/types";
+import { creationLobySchema } from "@/entities/lobby/types";
 import { createClient } from "@/shared/api/supabase/client";
 import { Button } from "@/shared/ui/button";
 import { Checkbox } from "@/shared/ui/checkbox";
