+++ src/_pages/lobby-detail/model/lobby-provider.tsx
@@ -10,7 +10,7 @@ import {
 } from "react";
 import { useLobbySubscription } from "@/_pages/lobby-detail/model/use-lobby-subscription";
 import { WelcomeForm } from "@/_pages/lobby-detail/ui/lobby-welcome-form";
-import type { Lobby, Member } from "@/features/lobby/types";
+import type { Lobby, Member } from "@/entities/lobby/types";
 
 type LobbyProviderProps = {
 	lobbyPromise: Promise<Lobby>;
