--- src/features/lobby/types/index.ts
@@ -1,19 +1,0 @@
-import type z from "zod";
-import type { creationLobySchema, lobbySchema } from "../schemas";
-
-export type Lobby = z.infer<typeof lobbySchema>;
-
-export type LobbyCreationDate = z.infer<typeof creationLobySchema>;
-
-export type Member = {
-	userId: string;
-	displayName: string;
-	userGoal: string;
-};
-
-export type Chat = {
-	chatId: string;
-	userId: string;
-	displayName: string;
-	content: string;
-};
